+++ server/authentication/utils.py
@@ -37,14 +37,3 @@ def reset_password(data: Dict[str, str]) -> None:
 def admin_login(password: str, email: str):
     password = sha256(password.encode()).hexdigest()
     return auth_model.admin_login(password=password, email=email)
-
-
-class RefreshToAccessPermission(BasePermission):
-    def has_permission(self, request, view):
-        token = get_token(headers=request.headers)
-        if issue_keys.is_valid_refresh(key=token) and not fire_watch.cache.sismember(
-            "Blacklist", token
-        ):
-            request.refresh_token = token
-            return True
-        return False
