+++ server/fire_watch/throttle.py
@@ -6,6 +6,6 @@ class Throttle(AnonRateThrottle):
     scope = "basic_throttle"
 
     def allow_request(self, request, view):
-        if request.method == "GET":
+        if request.method == "GET" and request.path != "auth/refresh":
             return True
         return super().allow_request(request, view)
